cosas

+++ random.hpp
@@ -6,7 +6,7 @@ using namespace std;
 
 long long int randomNo (int max, int min) {
     int *p = new int, x;
-    uintptr_t a = reinterpret_cast<uintptr_t> (p), b = reinterpret_cast<uintptr_t> (&x);
+    int a = reinterpret_cast<int> (p), b = reinterpret_cast<int> (&x);
     srand ((time(NULL) * a)/b);
     return (rand() % ((max-min)+1)) + min;
 }
